requirements.txt
streamlit-option-menu
streamlit
snowflake-connector-python